requirements.txt
annotated-types==0.7.0
faiss-cpu==1.9.0.post1
groq==0.13.1
langchain==0.3.14
langchain-groq==0.2.2
openai==1.59.3
python-dotenv==1.0.1
sentence-transformers==3.3.1
streamlit==1.41.1
typing_extensions==4.12.2
langgraph